Move clitest to tests/cli
