+++ tests/cli/gov_test.go
@@ -1,4 +1,4 @@
-package clitest
+package cli
 
 import (
 	"fmt"
